requirements.txt
git+https://github.com/korbinancell/remindmeofkorbinbot.git
discord.py==1.4.0
dlib==19.20.0
praw==7.1.0
opencv-python==4.3.0.36
load_dotenv